requirements.txt
# ===============================
# Core LangChain & AI
# ===============================
langchain
langchain-openai
langchain-anthropic
langchain-community

# ===============================
# LLM Providers
# ===============================
openai>=1.0.0
anthropic

# ===============================
# Vector Stores & Document Parsing
# ===============================
faiss-cpu
chromadb
pypdf
python-docx

# ===============================
# Web APIs & Backend
# ===============================
fastapi
uvicorn[standard]
requests
pydantic>=2.0.0

# ===============================
# Data Processing
# ===============================
pandas
numpy

# ===============================
# Jupyter & Interactive Dev
# ===============================
jupyter
jupyterlab
notebook
ipython
ipykernel
ipywidgets

# ===============================
# Testing
# ===============================
pytest
pytest-asyncio
pytest-cov

# ===============================
# Code Quality & Linting
# ===============================
black
flake8
mypy
pylint

# ===============================
# Utilities
# ===============================
python-dotenv
rich
tqdm
colorama

# ===============================
# Async & Networking
# ===============================
aiohttp

# ===============================
# Notebook Execution
# ===============================
nbclient
nbformat

# ===============================
# Additional
# ===============================
python-dateutil
pathlib